Added hello_world example, expanded appkit
This commit adds a Hello World program, which creates and displays a
minimalist Cocoa app with a Menu Bar, a Quit menu, and a Window.

It also wraps the Cocoa methods used in hello_world in static functions
in appkit, inside separate modules for different classes. Later on,
Cocoa methods could be implemented as traits on `id` if that turns out
to be easier.