Update core-graphics dependency to 0.2
This is what the rest of Servo is using -- so bump it here as well to
avoid test-tidy failure.

Note that this is indeed the version Servo is currently using core-text
with (as Servo still has a revision of core-text that didn't specify any
version constraints) -- so we are rather certain that this combination
works.