core-foundation: Use dep syntax for `with-uuid` feature.
This removes an implicit feature due to the optional dependency
`uuid`.

While this is user visible, any use of `uuid` as a feature previously
would not have enabled the support code and so it wouldn't have had
any benefit.

Bump the MSRV in CI to 1.64 (while only 1.60 is required here,
other features from 1.64 will be used).